os: boot using OpenSBI
The operating system now boots using the OpenSBI booting protocol but
lacks multithreading as that needs to be activated using the HSM
extension of the OpenSBI protocol.

The kernel also gets stuck after initializing the console as the UART
device is shared between the OpenSBI firmware and the kernel. Next step
should be to port the current console device to use the SBI debug
console extension.

OpenSBI does not always boot on hart 0 (mhartid = 0 passed through the
boot protocol). For that reason, the booting architecture of checking
the tp register does not work and two main functions have been created:
one for the booting hart, and another for the auxiliary. Other harts for
now don't really start as those need to be started via the HSM extension
of the SBI interface.