os: start other cores
Start all the cores at the same entry as the principal, the
infraestructure placed earlier makes it so that the secondary cores go
to main_secondary instead of main.

All cores are assumed to be contiguous, so we call sbi_hart_start until
we get an INVALID_PARAM, which signals an invalid hartid. Still, the
maximum supported cores is 8, as there stack buffer used by all cores is
allocated at compile time using the NCPU macro.