os: parse device tree header and verify magic number
The function for now will be called after initializing the console
(meaning that we assume the uart driver is the one implemented on the
source code). Still, it is nice to have printf available for debugging.

This in the future will need to be moved before initializing the
console, as parsing the device tree will result in using the correct
driver and correct pointers to memory mapped registers.