os: parse the /memory node in the DTB
Currently the reading of the reg property from the /memory node is kind
of custom to the device tree qemu gives us. The property should encode
64 bit integer pairs (first element being the start of the memory and
the second the size of the memory block). Instead qemu gives us garbage
in the first 64 bit value and next is the actual pair of values.

When decompiling the device tree into the device tree source, the actual
values of the reg property are (0x0, start, 0x0, size), which does not
make any sense.

This should be looked up in the future.