os: parse device tree for /chosen:stdout-path property
The stdout-path property of the /chosen node specifies a full path to
represent a device that could be useful for boot console output. We just
take advantage of that and check if the driver is compatible and fill in
the register address and interupt number.

The interrupt number should be taken with a grain of salt, as it does
not correctly follow the device tree conventions (interrupt tree stuff).
Still, as there are no nested interrupts in this machine, this should be
enough.