kernel: preserve PELP on user threads
When interruptions are taken during an indirect jump, the CPU state
machine that dictates if a landing pad is expected or not saves the
result in the PELP bit on the sstatus CSR register. That bit needs to be
preserved for user programs.

Its also important to 0 initialize the variable on the trapframe.