build: enable zicfiss extension on clang
For this to work the -fsanitize=shadow-call-stack flag needs to be used
with the -fcf-protection=return flag. Otherwise, without the second
flag, the software version is used instead.

Its important to only enable the shadow stack protection on user
threads, as using the sspush and sspopchk instructions without
configuring the shadow stack can result in illegal instruction
exceptions.