build: env target can connect to container
When opening multiple terminals, you had to manually connect to the
container (for example, in order to use gdb for debugging) using the
'docker exec' command and finding the container name that docker
generates automatically.

With this patch, if the container is already created, the env target
connects to the container automatically.