build: remove container when exiting
When exiting the container in the terminal the container stopped. When
trying to recreate it or attatch to it using the 'make env' command it
failed because the container already exists but can be attached to
because it's stopped.

This patch makes it so that when exiting the container is also
removed.