kernel: print process name on unexpected traps
When an unexpected trap occurs (any non syscall synchronous
interrupt), the kernel printed the scause register value and pid. This
can be a bit confusing as we can't really know the process that failed
only with those values.

With this patch the kernel also prints the process name stored in the
proc structure.