user: fix command line argument handling
The original implementation used undefined behavior in order to pass
the command line arguments from _main (wrapper that calls exit) to
main (the actual main function).

Instead of defining the functions with the correct arguments (inside
the runtime startup code), both are defined with empty parameters. As
the _main function only called the main function, the compiler did not
modify any of the values of the registers, thus automatically passing
the parameters from _main to main.

After adding the code to set up the software shadow stack, the
registers containing the argc and argv values are modified, passing
the wrong arguments into main. With this patch, the arguments are
taken explicitly and passed in the correct way.