kernel: fix unaligned segment loading on exec
This issue was causing weird parameter passing on strings with certain
configurations. See https://github.com/IkerGalardi/zeptOS/issues/39

Unaligned loads first loaded the unaligned part and then executed the
original aligned load algorithm. The issue was with the first
iteration of the aligned loads, as it always started on PGSIZE instead
of the the number of bytes written.

Fixes the before mentioned issue.